[BACKLOG-38429][Add] Credit Card Validator Plugin
