+++ engine/src/main/resources/org/pentaho/di/trans/step/messages/messages_en_US.properties
@@ -182,7 +182,6 @@ BaseStep.TypeTooltipDesc.ReplaceString=Replace all occurences a word in a string
 BaseStep.TypeLongDesc.MailInput=Email messages input
 BaseStep.TypeTooltipDesc.MailInput=Read POP3/IMAP server and retrieve messages
 BaseStep.TypeTooltipDesc.GetSubFolders=Read a parent folder and return all subfolders
-BaseStep.TypeLongDesc.CreditCardValidator=Credit card validator
 BaseStep.TypeLongDesc.RowFalttener=Row flattener
 BaseStep.TypeLongDesc.SelectValues=Select values
 BaseStep.TypeTooltipDesc.FileExists=Check if a file exists
@@ -287,7 +286,6 @@ BaseStep.TypeLongDesc.XMLOutput=XML output
 BaseStep.TypeTooltipDesc.Janino=Calculate the result of a Java Expression using Janino
 BaseStep.Log.StartingBuffersAllocation=Starting allocation of buffers & new threads...
 BaseStep.ColumnName.LinesSkipped=lines_skipped
-BaseStep.TypeTooltipDesc.CreditCardValidator=The Credit card validator step will help you tell\:\n(1) if a credit card number is valid (uses LUHN10 (MOD-10) algorithm)\n(2) which credit card vendor handles that number\n(VISA, MasterCard, Diners Club, EnRoute, American Express (AMEX),...)
 BaseStep.TypeTooltipDesc.SortedMerge=Sorted merge
 BaseStep.TypeLongDesc.SetValueField=Set field value
 BaseStepDialog.ClearAndAdd=C&lear and add all
