+++ plugins/credit-card-validator/impl/src/main/resources/org/pentaho/di/trans/steps/creditcardvalidator/messages/messages_en_US.properties
@@ -1,0 +1,59 @@
+#File generated by Hitachi Vantara Translator for package 'org.pentaho.di.trans.steps.CreditCardValidator' in locale 'en_US'
+#
+#
+#Thu Jun 19 19:10:46 CEST 2008
+BaseStep.TypeLongDesc.CreditCardValidator=Credit card validator
+BaseStep.TypeTooltipDesc.CreditCardValidator=The Credit card validator step will help you tell\:\n(1) if a credit card number is valid (uses LUHN10 (MOD-10) algorithm)\n(2) which credit card vendor handles that number\n(VISA, MasterCard, Diners Club, EnRoute, American Express (AMEX),...)
+CreditCardValidator.Error.ResultFieldMissing=Result fieldname is missing\!
+CreditCardValidatorMeta.CheckResult.ResultFieldOK=Result fieldname is provided
+CreditCardValidatorDialog.ResultField.Tooltip=Specify here the name of the result fieldname (Boolean)\nadded in the input stream 
+CreditCardValidatorDialog.Shell.Title=Credit card validator
+CreditCardValidatorDialog.FailedToGetFields.DialogMessage=Unable to get fields from previous steps because of an error
+CreditCardValidatorMeta.Exception.UnexpectedErrorReadingStepInfo=Unexpected error reading step information from the repository
+CreditCardValidatorDialog.Stepname.Label=Step name 
+CreditCardValidator.Log.UnexpectedError=Unexpected error in ''
+CreditCardValidatorDialog.CardType.Label=Credit card type field
+CreditCardValidatorDialog.FieldName.Label=Credit Card field 
+CreditCardValidatorDialog.FailedToGetFields.DialogTitle=Get fields failed
+CreditCardValidator.LineNumber=linenr {0}
+CreditCardValidator.Log.StartingToRun=Starting to run...
+CreditCardValidator.ErrorInStepRunning=Because of an error, this step can''t continue\: 
+CreditCardValidator.Log.ErrorFindingField=Error finding field\: 
+CreditCardValidatorMeta.Exception.UnableToSaveStepInfo=Unable to save step information to the repository for id_step\=
+CreditCardValidatorMeta.CheckResult.CouldNotReadFields=Couldn''t read fields from the previous step.
+CreditCardValidatorMeta.CheckResult.NoInpuReceived=No input received from other steps\!
+CreditCardValidator.Exception.CouldnotFindField=Couldn''t find field ''{0}'' in row\!
+CreditCardValidatorMeta.CheckResult.ReceivingInfoFromOtherSteps=Step is receiving info from other steps.
+CreditCardValidatorDialog.ResultField.Label=Result fieldname 
+CreditCardValidator.Error.CardFieldMissing=Card number fieldname is missing\!
+CreditCardValidatorMeta.CheckResult.ResultFieldMissing=Result fieldname is missing\!
+CreditCardValidatorDialog.CardType.Tooltip=Specify the name of the field that contains Credit Card type.\r\nThis field will be added to the input stream.
+CreditCardValidatorDialog.OutputFields.Label=Output fields
+CreditCardValidatorMeta.CheckResult.ErrorOccurred=An error occurred\: 
+CreditCardValidatorMeta.Exception.UnableToReadStepInfo=Unable to read step information from XML
+CreditCardValidatorMeta.CheckResult.CardFieldMissing=Credit Card number field name is missing\!
+CreditCardValidatorMeta.CheckResult.CardFieldOK=Credit Card number field name was specified
+CreditCardValidatorDialog.NotValidMsg.Label=Not valid Message
+CreditCardValidatorDialog.NotValidMsg.Tooltip=Not valid Message
+CreditCardValidator.Exception.ResultFieldMissing=Result fieldname is missing!
+CreditCardValidator.getOnlyDigits.Label=Get only digits?
+CreditCardValidator.getOnlyDigits.Tooltip=Extract only digits from credit card before checking.
+
+CreditCardValidator.Log.NotValidVisa=Not a valid number for Visa.
+CreditCardValidator.Log.NotValidMastercard=Not a valid number for Mastercard.
+CreditCardValidator.Log.NotValidAmericanExpress=Not a valid number for American Express.
+CreditCardValidator.Log.NotValidEnRoute=Not a valid number for En Route.
+CreditCardValidator.Log.NotValidDiners=Not a valid number for Diner's CLub/Carte Blanche.
+CreditCardValidator.Log.EmptyNumber=Card number cannot be blank!
+CreditCardValidator.Log.CardNotValid=This card is not valid!
+CreditCardValidator.Log.NotValidDiscover=Not a valid number for Discover
+CreditCardValidator.Log.NotValidJcb1=Not a valid number for JCB1
+CreditCardValidator.Log.NotValidJcb2=Not a valid number for JCB2
+CreditCardValidator.Log.NotValidMaestro=Not a valid number for BankCard
+CreditCardValidator.Log.NotValidMaestro=Not a valid number for MAESTRO
+CreditCardValidator.Log.NotValidSolo=Not a valid number for SOLO
+CreditCardValidator.Log.NotValidSwitch=Not a valid number for SWITCH
+CreditCardValidator.OnlyNumbers=Credit card number can only contain numbers
+CreditCardValidator.Log.NotValidAirplus=Not a valid AIRPLUS
+CreditCardValidator.Log.NotValidLaser=Not a valid LASER
+CreditCardValidator.Log.NotValidVoyager=Not a valid VOYAGER
