+++ plugins/credit-card-validator/impl/src/main/resources/org/pentaho/di/trans/steps/creditcardvalidator/messages/messages_es_AR.properties
@@ -1,0 +1,52 @@
+#File generated by Hitachi Vantara Translator for package 'org.pentaho.di.trans.steps.creditcardvalidator' in locale 'es_AR'
+#
+#
+#Mon Jun 28 16:02:10 ART 2010
+CreditCardValidatorDialog.Shell.Title=Validador de tarjetas de cr\u00E9dito
+CreditCardValidatorMeta.CheckResult.CardFieldOK=Se ha indicado el campo para el n\u00FAmero de tarjeta de cr\u00E9dito
+CreditCardValidator.Error.ResultFieldMissing=Falta el nombre del campo para el resultado.
+CreditCardValidator.OnlyNumbers=El n\u00FAmero de tarjeta de cr\u00E9dito s\u00F3lo puede contener n\u00FAmeros
+CreditCardValidator.Log.NotValidMastercard=No es un n\u00FAmero v\u00E1lido para MasterCard.
+CreditCardValidator.Log.NotValidDiscover=No es un n\u00FAmero v\u00E1lido para Discover
+CreditCardValidatorMeta.CheckResult.ResultFieldMissing=Falta el campo para el resultado.
+CreditCardValidator.Log.EmptyNumber=El n\u00FAmero de tarjeta no puede estar vac\u00EDo
+CreditCardValidator.Log.NotValidMaestro=No es un n\u00FAmero v\u00E1lido para MAESTRO
+CreditCardValidator.Exception.ResultFieldMissing=Falta el nombre de campo para el resultado.
+CreditCardValidatorDialog.CardType.Tooltip=Indique el nombre del campo que contiene el tipo de tarjeta de cr\u00E9dito.\nEste campo ser\u00E1 a\u00F1adido al flujo de datos.
+CreditCardValidatorDialog.ResultField.Tooltip=Indique aqu\u00ED el nombre del campo para el resultado (booleano)\n a a\u00F1adir en el flujo
+CreditCardValidatorMeta.CheckResult.ReceivingInfoFromOtherSteps=El paso est\u00E1 recibiendo informaci\u00F3n de otros pasos.
+CreditCardValidatorMeta.Exception.UnableToReadStepInfo=Imposible escribir informaci\u0144 del paso desde XML
+CreditCardValidator.Log.NotValidVoyager=No es un n\u00FAmero v\u00E1lido para VOYAGER
+CreditCardValidator.Exception.CouldnotFindField=No se encontr\u00F3 al campo "{0}" en la fila.
+CreditCardValidatorDialog.OutputFields.Label=Campos de salida
+CreditCardValidatorMeta.Exception.UnexpectedErrorReadingStepInfo=Error inesperado al leer informaci\u00F3n del paso desde el repositorio
+CreditCardValidator.Log.NotValidJcb2=No es un n\u00FAmero v\u00E1lido para JCB2
+CreditCardValidator.Log.NotValidJcb1=No es un n\u00FAmero v\u00E1lido para JCB1
+CreditCardValidator.Log.NotValidAirplus=No es un n\u00FAmero v\u00E1lido para AIRPLUS
+CreditCardValidator.Log.StartingToRun=Comenzando la ejecuci\u00F3n...
+CreditCardValidatorDialog.FailedToGetFields.DialogMessage=Imposible obtener campos de pasos previos debido a un error
+CreditCardValidatorDialog.FailedToGetFields.DialogTitle=Fall\u00F3 la obtenci\u00F3n de campos
+CreditCardValidator.getOnlyDigits.Tooltip=Extraer s\u00F3lo los d\u00EDgitos de la tarjeta de cr\u00E9dito antes de verificar.
+CreditCardValidator.Log.UnexpectedError=Error inesperado en ''
+CreditCardValidator.Log.CardNotValid=Esta tarjeta no es v\u00E1lida.
+CreditCardValidatorDialog.NotValidMsg.Label=No es un mensaje v\u00E1lido
+CreditCardValidatorDialog.ResultField.Label=Nombre de campo para resultado
+CreditCardValidator.Log.NotValidVisa=No es un n\u00FAmero v\u00E1lido para Visa
+CreditCardValidator.LineNumber=NroLinea {0}
+CreditCardValidatorDialog.FieldName.Label=Campo de tarjeta de cr\u00E9dito
+CreditCardValidator.Log.NotValidDiners=No es un n\u00FAmero v\u00E1lido para Diners Club/Carte Blanche.
+CreditCardValidatorDialog.CardType.Label=Campo para tipo de tarjeta de cr\u00E9dito
+CreditCardValidator.ErrorInStepRunning=Este paso no puede continuar debido a un error.
+CreditCardValidator.Log.NotValidSwitch=No es un n\u00FAmero v\u00E1lido para SWITCH
+CreditCardValidator.Log.NotValidAmericanExpress=No es un n\u00FAmero v\u00E1lido para American Express.
+CreditCardValidatorDialog.Stepname.Label=Nombre de paso
+CreditCardValidator.Log.NotValidSolo=No es un n\u00FAmero v\u00E1lido para SOLO
+CreditCardValidator.Log.NotValidEnRoute=No es un n\u00FAmero v\u00E1lido para En Route.
+CreditCardValidator.getOnlyDigits.Label=\u00BFObtener s\u00F3lo los d\u00EDgitos?
+CreditCardValidatorMeta.CheckResult.CardFieldMissing=Falta el campo para el n\u00FAmero de tarjeta de cr\u00E9dito.
+CreditCardValidatorDialog.NotValidMsg.Tooltip=No es un mensaje v\u00E1lido
+CreditCardValidatorMeta.Exception.UnableToSaveStepInfo=Imposible guardar informaci\u00F3n del paso en el repositorio para el paso con ID\=
+CreditCardValidatorMeta.CheckResult.NoInpuReceived=No se ha recibido entrada de otros pasos.
+CreditCardValidator.Error.CardFieldMissing=Falta el nombre del campo para el n\u00FAmero de tarjeta.
+CreditCardValidatorMeta.CheckResult.ResultFieldOK=Se ha encontrado el campo para el resultado
+CreditCardValidator.Log.NotValidLaser=No es un n\u00FAmero v\u00E1lido para LASER
