+++ plugins/credit-card-validator/impl/src/main/resources/org/pentaho/di/trans/steps/creditcardvalidator/messages/messages_fr_FR.properties
@@ -1,0 +1,56 @@
+#File generated by Hitachi Vantara Translator for package 'org.pentaho.di.trans.steps.CreditCardValidator' in locale 'en_US'
+#
+#
+#Thu Jun 19 19:10:46 CEST 2008
+CreditCardValidator.Error.ResultFieldMissing=Le champ r?sultat n''a pas ?t? renseign?\!
+CreditCardValidatorMeta.CheckResult.ResultFieldOK=Le champ r?sultat a ?t? sp?cifi?.
+CreditCardValidatorDialog.ResultField.Tooltip=Indiquez dans cette zone, le nom du champ contenant\nle r?sultat de la validation (Binaire).\nCe champ sera ajout? au flux d''entr?e. 
+CreditCardValidatorDialog.Shell.Title=Validation carte de cr?dit
+CreditCardValidatorDialog.FailedToGetFields.DialogMessage=Impossible de r?cup?rer les champs depuis les ?tapes pr?c?dentes
+CreditCardValidatorMeta.Exception.UnexpectedErrorReadingStepInfo=PDI a rencontr? une erreur lors de la lecture des informations depuis le r?f?rentiel!
+CreditCardValidatorDialog.Stepname.Label=Nom ?tape
+CreditCardValidator.Log.UnexpectedError=Erreur inattendue
+CreditCardValidatorDialog.CardType.Label=Champ type carte
+CreditCardValidatorDialog.FieldName.Label=Champ N?carte 
+CreditCardValidatorDialog.FailedToGetFields.DialogTitle=R?cup?rer champs
+CreditCardValidator.LineNumber=N?ligne {0}
+CreditCardValidator.Log.StartingToRun=D?marrage...
+CreditCardValidator.ErrorInStepRunning=Erreur inattendue! Arr?t de l''ex?cution\: 
+CreditCardValidator.Log.ErrorFindingField=Erreur champ\: 
+CreditCardValidatorMeta.Exception.UnableToSaveStepInfo=Impossible de sauvegarder dans le r?f?rentiel, les informations de l''?tape identifi? par l''id\=
+CreditCardValidatorMeta.CheckResult.CouldNotReadFields=Impossible de r?cup?rer les champs depuis les ?tapes pr?c?dentes.
+CreditCardValidatorMeta.CheckResult.NoInpuReceived=Aucun champ n''a ?t? r?cup?r? depuis les autres ?tapes\!
+CreditCardValidator.Exception.CouldnotFindField=Impossible de trouver le champ ''{0}'' dans le flux d''entr?e\!
+CreditCardValidatorMeta.CheckResult.ReceivingInfoFromOtherSteps=L''?tape r?cup?re des informations depuis les autres ?tapes.
+CreditCardValidatorDialog.ResultField.Label=Nom champ r?sultat 
+CreditCardValidator.Error.CardFieldMissing=Le champ contenant le num?ro de carte n''a pas ?t? renseign?\!
+CreditCardValidatorMeta.CheckResult.ResultFieldMissing=Le champ contenant le r?sulat de la validation n''a pas ?t? renseign?\!
+CreditCardValidatorDialog.CardType.Tooltip=Indiquez dans cette zone le nom du champ qui contiendra le type de carte.\nCe champ (cha?ne de caract?res) sera ajout? au flux d''entr?e.
+CreditCardValidatorDialog.OutputFields.Label=Champs en sortie
+CreditCardValidatorMeta.CheckResult.ErrorOccurred=PDI a rencontr? une erreur\: 
+CreditCardValidatorMeta.Exception.UnableToReadStepInfo=Impossible de lire les informations de l''?tape depuis le fichier XML!
+CreditCardValidatorMeta.CheckResult.CardFieldMissing=Le champ contenant le num?ro de carte n''a pas ?t? renseign?\!
+CreditCardValidatorMeta.CheckResult.CardFieldOK=Le champ contenant le num?ro de carte a ?t? renseign?\.
+CreditCardValidatorDialog.NotValidMsg.Label=Champ message invalide
+CreditCardValidatorDialog.NotValidMsg.Tooltip=Saisissez dans cette zone le nom du champ\nqui contiendra les messages renvoy?s lors du process de validation.\nCe champ (cha?ne de caract?res) sera ajout? au flux d''entr?e.
+CreditCardValidator.Exception.ResultFieldMissing=Le champ contenant le r?sultat n''a pas ?t? renseign?!
+CreditCardValidator.getOnlyDigits.Label=Extraire nombres
+CreditCardValidator.getOnlyDigits.Tooltip=S?lectionnez cette option si vous souhaitez\nextraire les nombres depuis le num?ro de la carte.\nCette extraction sera r?alis?e avant le processus de validation de la carte.
+
+CreditCardValidator.Log.NotValidVisa=Cette carte n''est pas une carte Visa valide.
+CreditCardValidator.Log.NotValidMastercard=Cette carte n''est pas une carte Mastercard valide.
+CreditCardValidator.Log.NotValidAmericanExpress=Cette carte n''est pas une carte American Express valide.
+CreditCardValidator.Log.NotValidEnRoute=Cette carte n''est pas une carte En Route valide.
+CreditCardValidator.Log.NotValidDiners=Cette carte n''est pas une carte CLub/Carte Blanche valide.
+CreditCardValidator.Log.EmptyNumber=Le num?ro de carte est vide!
+CreditCardValidator.Log.CardNotValid=Cette carte est invalide!
+CreditCardValidator.Log.NotValidDiscover=Cette carte n''est pas une carte Discover valide.
+CreditCardValidator.Log.NotValidJcb1=Cette carte n''est pas une carte Jcb1 valide.
+CreditCardValidator.Log.NotValidJcb2=Cette carte n''est pas une carte Jcb2 valide.
+CreditCardValidator.Log.NotValidMaestro=Cette carte n''est pas une carte Maestro valide.
+CreditCardValidator.Log.NotValidSolo=Cette carte n''est pas une carte Solo valide.
+CreditCardValidator.Log.NotValidSwitch=Cette carte n''est pas une carte Switch valide.
+CreditCardValidator.OnlyNumbers=Les num?ros de carte ne doivent contenir que des chiffres
+CreditCardValidator.Log.NotValidAirplus=Cette carte n''est pas une carte Airplus valide.
+CreditCardValidator.Log.NotValidLaser=Cette carte n''est pas une carte Laser valide.
+CreditCardValidator.Log.NotValidVoyager=Cette carte n''est pas une carte Voyager valide.
