+++ plugins/credit-card-validator/impl/src/main/resources/org/pentaho/di/trans/steps/creditcardvalidator/messages/messages_it_IT.properties
@@ -1,0 +1,52 @@
+#File generated by Hitachi Vantara Translator for package 'org.pentaho.di.trans.steps.creditcardvalidator' in locale 'it_IT'
+#
+#
+#Sun Jun 27 23:49:35 CEST 2010
+CreditCardValidatorDialog.Shell.Title=Validatore carta di credito
+CreditCardValidatorMeta.CheckResult.CardFieldOK=Il campo col numero della carta di credito \u00E8 stato specificato
+CreditCardValidator.Error.ResultFieldMissing=Manca il nome campo del risultato\!
+CreditCardValidator.OnlyNumbers=Un numero di carta di credito pu\u00F2 contenere solo numeri
+CreditCardValidator.Log.NotValidMastercard=Non \u00E8 un numero valido per Mastercard.
+CreditCardValidator.Log.NotValidDiscover=Numero non valido per Discover
+CreditCardValidatorMeta.CheckResult.ResultFieldMissing=Manca il nome campo del risultato\!
+CreditCardValidator.Log.EmptyNumber=Il numero della carta non pu\u00F2 restare bianco\!
+CreditCardValidator.Log.NotValidMaestro=Numero non valido per MAESTRO
+CreditCardValidator.Exception.ResultFieldMissing=Manca il nome campo del risultato\!
+CreditCardValidatorDialog.CardType.Tooltip=Specificare qua il nome del campo che contiene il tipo di carta di credito.\nQuesto campo verr\u00E0 aggiunto all''input stream.
+CreditCardValidatorDialog.ResultField.Tooltip=Specificare qua il nome del campo risultato aggiunto nell''input stream.
+CreditCardValidatorMeta.CheckResult.ReceivingInfoFromOtherSteps=Il passo sta ricevendo informazioni dagli altri passi.
+CreditCardValidatorMeta.Exception.UnableToReadStepInfo=Impossibile leggere le informazioni del passo da XML
+CreditCardValidator.Log.NotValidVoyager=Non \u00E8 un VOYAGER valido
+CreditCardValidator.Exception.CouldnotFindField=Impossibile trovare il campo ''{0}'' nella riga\!
+CreditCardValidatorDialog.OutputFields.Label=Campi di output
+CreditCardValidatorMeta.Exception.UnexpectedErrorReadingStepInfo=Errore inatteso durante la lettura delle informazioni del passo dal repository
+CreditCardValidator.Log.NotValidJcb2=Numero non valido per JCB2
+CreditCardValidator.Log.NotValidJcb1=Numero non valido per JCB1
+CreditCardValidator.Log.NotValidAirplus=Non \u00E8 un AIRPLUS valido
+CreditCardValidator.Log.StartingToRun=Avvio dell''esecuzione...
+CreditCardValidatorDialog.FailedToGetFields.DialogMessage=Impossibile prelevare i campi dai passi precedenti a causa di un errore
+CreditCardValidatorDialog.FailedToGetFields.DialogTitle=Prelievo campi fallito
+CreditCardValidator.getOnlyDigits.Tooltip=Estrae solo le cifre dalla carta di credito prima di controllare.
+CreditCardValidator.Log.UnexpectedError=Errore inatteso in ''
+CreditCardValidator.Log.CardNotValid=Questa carta non \u00E8 valida\!
+CreditCardValidatorDialog.NotValidMsg.Label=Non \u00E8 un messaggio valido
+CreditCardValidatorDialog.ResultField.Label=Nome campo del risultato
+CreditCardValidator.Log.NotValidVisa=Non \u00E8 un numero valido per Visa.
+CreditCardValidator.LineNumber=linea n\u00B0 {0}
+CreditCardValidatorDialog.FieldName.Label=Campo carta di credito
+CreditCardValidator.Log.NotValidDiners=Non \u00E8 un numero valido per Diner's CLub/Carte Blanche.
+CreditCardValidatorDialog.CardType.Label=Tipo campo carta di credito
+CreditCardValidator.ErrorInStepRunning=A causa di un errore, questo passo non pu\u00F2 continuare\: 
+CreditCardValidator.Log.NotValidSwitch=Numero non valido per SWITCH
+CreditCardValidator.Log.NotValidAmericanExpress=Non \u00E8 un numero valido per American Express.
+CreditCardValidatorDialog.Stepname.Label=Nome del passo
+CreditCardValidator.Log.NotValidSolo=Numero non valido per SOLO
+CreditCardValidator.Log.NotValidEnRoute=Non \u00E8 un numero valido per En Route.
+CreditCardValidator.getOnlyDigits.Label=Ricevere solo le cifre?
+CreditCardValidatorMeta.CheckResult.CardFieldMissing=Manca il nome del campo con il numero della carta di credito\!
+CreditCardValidatorDialog.NotValidMsg.Tooltip=Non \u00E8 un messaggio valido
+CreditCardValidatorMeta.Exception.UnableToSaveStepInfo=Impossibile salvare le informazioni del passo nel repository per id_step\=
+CreditCardValidatorMeta.CheckResult.NoInpuReceived=Nessun input ricevuto dagli altri passi\!
+CreditCardValidator.Error.CardFieldMissing=Manca il nome campo con il numero della carta\!
+CreditCardValidatorMeta.CheckResult.ResultFieldOK=Il nome campo del risultato \u00E8 stato fornito
+CreditCardValidator.Log.NotValidLaser=Non \u00E8 un LASER valido
