+++ plugins/credit-card-validator/impl/src/main/resources/org/pentaho/di/trans/steps/creditcardvalidator/messages/messages_ja_JP.properties
@@ -1,0 +1,56 @@
+#File generated by Hitachi Vantara Translator for package 'org.pentaho.di.trans.steps.creditcardvalidator' in locale 'ja_JP'
+#
+#
+#Sat Sep 19 14:22:26 JST 2009
+CreditCardValidator.Error.ResultFieldMissing=Result fieldname is missing\!
+CreditCardValidatorMeta.CheckResult.ResultFieldOK=Result fieldname is provided
+CreditCardValidatorDialog.ResultField.Tooltip=\u5165\u529b\u30b9\u30c8\u30ea\u30fc\u30e0\u5185\u3067\u8ffd\u52a0\u3055\u308c\u305f\u7d50\u679c\u30d5\u30a3\u30fc\u30eb\u30c9\u540d(Boolean)\u306e\u540d\u524d\u3092\u6307\u5b9a\u3059\u308b\u5834\u5408\u306f\u30c1\u30a7\u30c3\u30af\u30dc\u30c3\u30af\u30b9\u3092\u6709\u52b9\u306b\u3057\u3066\u304f\u3060\u3055\u3044\u3002
+CreditCardValidatorDialog.Shell.Title=\u30af\u30ec\u30b8\u30c3\u30c8\u30ab\u30fc\u30c9\u691c\u8a3c
+CreditCardValidatorDialog.FailedToGetFields.DialogMessage=\u30a8\u30e9\u30fc\u306e\u70ba\u3001\u5148\u884c\u30b9\u30c6\u30c3\u30d7\u304b\u3089\u30d5\u30a3\u30fc\u30eb\u30c9\u3092\u53d6\u5f97\u51fa\u6765\u307e\u305b\u3093\u3002
+CreditCardValidatorMeta.Exception.UnexpectedErrorReadingStepInfo=Unexpected error reading step information from the repository
+CreditCardValidatorDialog.Stepname.Label=\u30b9\u30c6\u30c3\u30d7\u540d 
+CreditCardValidator.Log.UnexpectedError=Unexpected error in ''
+CreditCardValidatorDialog.CardType.Label=\u30ab\u30fc\u30c9\u30bf\u30a4\u30d7
+CreditCardValidatorDialog.FieldName.Label=\u30d5\u30a3\u30fc\u30eb\u30c9\u540d
+CreditCardValidatorDialog.FailedToGetFields.DialogTitle=\u30d5\u30a3\u30fc\u30eb\u30c9\u53d6\u5f97\u5931\u6557
+CreditCardValidator.LineNumber=linenr {0}
+CreditCardValidator.Log.StartingToRun=Starting to run...
+CreditCardValidator.ErrorInStepRunning=Because of an error, this step can''t continue\: 
+CreditCardValidator.Log.ErrorFindingField=Error finding field\: 
+CreditCardValidatorMeta.Exception.UnableToSaveStepInfo=Unable to save step information to the repository for id_step\=
+CreditCardValidatorMeta.CheckResult.CouldNotReadFields=Couldn''t read fields from the previous step.
+CreditCardValidatorMeta.CheckResult.NoInpuReceived=No input received from other steps\!
+CreditCardValidator.Exception.CouldnotFindField=Couldn''t find field ''{0}'' in row\!
+CreditCardValidatorMeta.CheckResult.ReceivingInfoFromOtherSteps=Step is receiving info from other steps.
+CreditCardValidatorDialog.ResultField.Label=\u7d50\u679c
+CreditCardValidator.Error.CardFieldMissing=Card number fieldname is missing\!
+CreditCardValidatorMeta.CheckResult.ResultFieldMissing=Result fieldname is missing\!
+CreditCardValidatorDialog.CardType.Tooltip=\u30af\u30ec\u30b8\u30c3\u30c8\u30ab\u30fc\u30c9\u30bf\u30a4\u30d7\u3092\u542b\u3080\u30d5\u30a3\u30fc\u30eb\u30c9\u540d\u3092\u6307\u5b9a\u3059\u308b\u5834\u5408\u306f\u30c1\u30a7\u30c3\u30af\u30dc\u30c3\u30af\u30b9\u3092\u6709\u52b9\u306b\u3057\u3066\u304f\u3060\u3055\u3044\u3002
+CreditCardValidatorDialog.OutputFields.Label=\u51fa\u529b\u30d5\u30a3\u30fc\u30eb\u30c9
+CreditCardValidatorMeta.CheckResult.ErrorOccurred=An error occurred\: 
+CreditCardValidatorMeta.Exception.UnableToReadStepInfo=Unable to read step information from XML
+CreditCardValidatorMeta.CheckResult.CardFieldMissing=Credit Card number field name is missing\!
+CreditCardValidatorMeta.CheckResult.CardFieldOK=Credit Card number field name was specified
+CreditCardValidatorDialog.NotValidMsg.Label=\u30a8\u30e9\u30fc\u30e1\u30c3\u30bb\u30fc\u30b8
+CreditCardValidatorDialog.NotValidMsg.Tooltip=\u6709\u52b9\u306a\u30e1\u30c3\u30bb\u30fc\u30b8\u3067\u306f\u3042\u308a\u307e\u305b\u3093\u3002
+CreditCardValidator.Exception.ResultFieldMissing=Result fieldname is missing!
+CreditCardValidator.getOnlyDigits.Label=\u6570\u5b57\u306e\u307f\u53d6\u5f97
+CreditCardValidator.getOnlyDigits.Tooltip=\u78ba\u8a8d\u306e\u524d\u306b\u30af\u30ec\u30b8\u30c3\u30c8\u30ab\u30fc\u30c9\u30d5\u30a3\u30fc\u30eb\u30c9\u304b\u3089\u6570\u5b57\u306e\u307f\u3092\u62bd\u51fa\u3059\u308b\u5834\u5408\u306f\u30c1\u30a7\u30c3\u30af\u30dc\u30c3\u30af\u30b9\u3092\u6709\u52b9\u306b\u3057\u3066\u304f\u3060\u3055\u3044\u3002
+
+CreditCardValidator.Log.NotValidVisa=Not a valid number for Visa.
+CreditCardValidator.Log.NotValidMastercard=Not a valid number for Mastercard.
+CreditCardValidator.Log.NotValidAmericanExpress=Not a valid number for American Express.
+CreditCardValidator.Log.NotValidEnRoute=Not a valid number for En Route.
+CreditCardValidator.Log.NotValidDiners=Not a valid number for Diner's CLub/Carte Blanche.
+CreditCardValidator.Log.EmptyNumber=Card number cannot be blank!
+CreditCardValidator.Log.CardNotValid=This card is not valid!
+CreditCardValidator.Log.NotValidDiscover=Not a valid number for Discover
+CreditCardValidator.Log.NotValidJcb1=Not a valid number for JCB1
+CreditCardValidator.Log.NotValidJcb2=Not a valid number for JCB2
+CreditCardValidator.Log.NotValidMaestro=Not a valid number for MAESTRO
+CreditCardValidator.Log.NotValidSolo=Not a valid number for SOLO
+CreditCardValidator.Log.NotValidSwitch=Not a valid number for SWITCH
+CreditCardValidator.OnlyNumbers=Credit card number can only contain numbers
+CreditCardValidator.Log.NotValidAirplus=Not a valid AIRPLUS
+CreditCardValidator.Log.NotValidLaser=Not a valid LASER
+CreditCardValidator.Log.NotValidVoyager=Not a valid VOYAGER
