+++ plugins/credit-card-validator/impl/src/main/resources/org/pentaho/di/trans/steps/creditcardvalidator/messages/messages_ko_KR.properties
@@ -1,0 +1,49 @@
+#File generated by Hitachi Vantara Translator for package 'org.pentaho.di.trans.steps.creditcardvalidator' in locale 'ko_KR'
+#
+#
+#Mon Dec 21 11:34:41 KST 2009
+CreditCardValidatorDialog.Shell.Title=Credit card validator
+CreditCardValidatorMeta.CheckResult.CardFieldOK=\uC2E0\uC6A9\uCE74\uB4DC \uBC88\uD638 \uD544\uB4DC \uC774\uB984\uC744 \uC9C0\uC815\uD558\uC600\uC2B5\uB2C8\uB2E4
+CreditCardValidator.Error.ResultFieldMissing=\uACB0\uACFC \uD544\uB4DC\uC774\uB984\uC774 \uC5C6\uC2B5\uB2C8\uB2E4\!
+CreditCardValidator.OnlyNumbers=\uC2E0\uC6A9\uCE74\uB4DC \uBC88\uD638\uB294 \uC22B\uC790\uB9CC \uD3EC\uD568\uD560 \uC218 \uC788\uC2B5\uB2C8\uB2E4
+CreditCardValidator.Log.NotValidMastercard=\uC720\uD6A8\uD55C Mastercard \uCE74\uB4DC\uBC88\uD638\uAC00 \uC544\uB2D9\uB2C8\uB2E4.
+CreditCardValidator.Log.NotValidDiscover=\uC720\uD6A8\uD55C Discover \uBC88\uD638\uAC00 \uC544\uB2D9\uB2C8\uB2E4
+CreditCardValidatorMeta.CheckResult.ResultFieldMissing=\uACB0\uACFC \uD544\uB4DC\uC774\uB984\uC774 \uC5C6\uC2B5\uB2C8\uB2E4\!
+CreditCardValidator.Log.EmptyNumber=\uCE74\uB4DC \uBC88\uD638\uB97C \uC785\uB825\uD558\uC2ED\uC2DC\uC624\!
+CreditCardValidator.Log.NotValidMaestro=\uC720\uD6A8\uD55C MAESTRO \uBC88\uD638\uAC00 \uC544\uB2D9\uB2C8\uB2E4
+CreditCardValidator.Exception.ResultFieldMissing=\uACB0\uACFC \uD544\uB4DC\uC774\uB984\uC774 \uC5C6\uC2B5\uB2C8\uB2E4\!
+CreditCardValidatorDialog.CardType.Tooltip=\uC2E0\uC6A9\uCE74\uB4DC \uC885\uB958\uB97C \uD3EC\uD568\uD55C \uD544\uB4DC\uC758 \uC774\uB984\uC744 \uC9C0\uC815\uD558\uC2ED\uC2DC\uC624.\r\n\uC774 \uD544\uB4DC\uB294 \uC785\uB825 \uC2A4\uD2B8\uB9BC\uC5D0 \uD3EC\uD568\uB420 \uAC83 \uC785\uB2C8\uB2E4.
+CreditCardValidatorDialog.ResultField.Tooltip=\uC5EC\uAE30\uC5D0 \uACB0\uACFC \uD544\uB4DC\uC758 \uC774\uB984 \uC9C0\uC815 (Boolean)\r\n\uC785\uB825 \uC2A4\uD2B8\uB9BC\uC5D0\uC11C \uCD94\uAC00\uB428
+CreditCardValidatorMeta.CheckResult.ReceivingInfoFromOtherSteps=Step\uC774 \uB2E4\uB978 Step\uC5D0\uC11C \uC815\uBCF4\uB97C \uBC1B\uC558\uC2B5\uB2C8\uB2E4.
+CreditCardValidatorMeta.Exception.UnableToReadStepInfo=XML\uC5D0\uC11C Step \uC815\uBCF4\uB97C \uC77D\uC744 \uC218 \uC5C6\uC2B5\uB2C8\uB2E4
+CreditCardValidator.Exception.CouldnotFindField=\uB85C\uC6B0\uC5D0\uC11C \uD544\uB4DC ''{0}''\uC744 \uCC3E\uC744 \uC218 \uC5C6\uC2B5\uB2C8\uB2E4\!
+CreditCardValidatorDialog.OutputFields.Label=\uCD9C\uB825 \uD544\uB4DC
+CreditCardValidatorMeta.Exception.UnexpectedErrorReadingStepInfo=\uC800\uC7A5\uC18C\uC5D0\uC11C Step \uC815\uBCF4 \uC77D\uAE30 \uC624\uB958
+CreditCardValidator.Log.NotValidJcb2=\uC720\uD6A8\uD55C JCB2 \uBC88\uD638\uAC00 \uC544\uB2D9\uB2C8\uB2E4
+CreditCardValidator.Log.NotValidJcb1=\uC720\uD6A8\uD55C JCB1 \uBC88\uD638\uAC00 \uC544\uB2D9\uB2C8\uB2E4
+CreditCardValidator.Log.StartingToRun=\uC2E4\uD589 \uC2DC\uC791...
+CreditCardValidatorDialog.FailedToGetFields.DialogMessage=\uC624\uB958\uB54C\uBB38\uC5D0 \uC774\uC804 Step\uC5D0\uC11C \uD544\uB4DC\uB97C \uAC00\uC838\uC62C \uC218 \uC5C6\uC2B5\uB2C8\uB2E4
+CreditCardValidatorDialog.FailedToGetFields.DialogTitle=\uD544\uB4DC \uAC00\uC838\uC624\uAE30 \uC2E4\uD328
+CreditCardValidator.getOnlyDigits.Tooltip=\uD655\uC778 \uC804\uC5D0 \uC2E0\uC6A9\uCE74\uB4DC\uC5D0\uC11C \uC22B\uC790\uB9CC \uCD94\uCD9C
+CreditCardValidator.Log.UnexpectedError=\uC54C \uC218 \uC5C6\uB294 \uC624\uB958 ''
+CreditCardValidator.Log.CardNotValid=\uC720\uD6A8\uD55C \uCE74\uB4DC\uAC00 \uC544\uB2D9\uB2C8\uB2E4\!
+CreditCardValidatorDialog.NotValidMsg.Label=\uC720\uD6A8\uD558\uC9C0 \uC54A\uC740 \uACBD\uC6B0 \uBA54\uC138\uC9C0
+CreditCardValidatorDialog.ResultField.Label=\uACB0\uACFC \uD544\uB4DC\uC774\uB984
+CreditCardValidator.Log.NotValidVisa=\uC720\uD6A8\uD55C Visa \uCE74\uB4DC\uBC88\uD638\uAC00 \uC544\uB2D9\uB2C8\uB2E4.
+CreditCardValidator.LineNumber=\uC904\uBC88\uD638 {0}
+CreditCardValidatorDialog.FieldName.Label=\uC2E0\uC6A9\uCE74\uB4DC \uD544\uB4DC
+CreditCardValidator.Log.NotValidDiners=\uC720\uD6A8\uD55C Diner's CLub/Carte Blanche \uBC88\uD638\uAC00 \uC544\uB2D9\uB2C8\uB2E4.
+CreditCardValidatorDialog.CardType.Label=\uC2E0\uC6A9\uCE74\uB4DC \uC885\uB958 \uD544\uB4DC
+CreditCardValidator.ErrorInStepRunning=\uC624\uB958 \uB54C\uBB38\uC5D0, Step\uC744 \uACC4\uC18D\uD560 \uC218 \uC5C6\uC2B5\uB2C8\uB2E4\:
+CreditCardValidator.Log.NotValidSwitch=\uC720\uD6A8\uD55C SWITCH \uBC88\uD638\uAC00 \uC544\uB2D9\uB2C8\uB2E4.
+CreditCardValidator.Log.NotValidAmericanExpress=\uC720\uD6A8\uD55C American Express \uBC88\uD638\uAC00 \uC544\uB2D9\uB2C8\uB2E4.
+CreditCardValidatorDialog.Stepname.Label=Step \uC774\uB984 
+CreditCardValidator.Log.NotValidSolo=\uC720\uD6A8\uD55C SOLO \uBC88\uD638\uAC00 \uC544\uB2D9\uB2C8\uB2E4.
+CreditCardValidator.Log.NotValidEnRoute=\uC720\uD6A8\uD55C En Route \uCE74\uB4DC\uBC88\uD638\uAC00 \uC544\uB2D9\uB2C8\uB2E4.
+CreditCardValidator.getOnlyDigits.Label=\uC22B\uC790\uB9CC \uAC00\uC838\uC624\uAE30?
+CreditCardValidatorMeta.CheckResult.CardFieldMissing=\uC2E0\uC6A9\uCE74\uB4DC \uBC88\uD638 \uD544\uB4DC \uC774\uB984\uC744 \uC9C0\uC815\uD558\uC2ED\uC2DC\uC624\!
+CreditCardValidatorDialog.NotValidMsg.Tooltip=\uC720\uD6A8\uD55C \uBA54\uC138\uC9C0\uAC00 \uC544\uB2D9\uB2C8\uB2E4
+CreditCardValidatorMeta.Exception.UnableToSaveStepInfo=\uC800\uC7A5\uC18C\uC5D0 Step \uC815\uBCF4\uB97C \uC800\uC7A5\uD560 \uC218 \uC5C6\uC2B5\uB2C8\uB2E4. id_step\=
+CreditCardValidatorMeta.CheckResult.NoInpuReceived=\uB2E4\uB978 Step\uC5D0\uC11C \uBC1B\uC740 \uC785\uB825\uC774 \uC5C6\uC2B5\uB2C8\uB2E4\!
+CreditCardValidator.Error.CardFieldMissing=\uCE74\uB4DC \uBC88\uD638 \uD544\uB4DC\uC774\uB984\uC774 \uC5C6\uC2B5\uB2C8\uB2E4\!
+CreditCardValidatorMeta.CheckResult.ResultFieldOK=\uACB0\uACFC \uD544\uB4DC\uC774\uB984\uC774 \uC788\uC2B5\uB2C8\uB2E4
