+++ plugins/credit-card-validator/impl/src/test/java/org/pentaho/di/trans/steps/creditcardvalidator/CreditCardVerifierTest.java
@@ -2,7 +2,7 @@
  *
  * Pentaho Data Integration
  *
- * Copyright (C) 2002-2017 by Hitachi Vantara : http://www.pentaho.com
+ * Copyright (C) 2002-2023 by Hitachi Vantara : http://www.pentaho.com
  *
  *******************************************************************************
  *
