+++ plugins/credit-card-validator/ui/src/main/java/org/pentaho/di/ui/trans/steps/creditcardvalidator/CreditCardValidatorDialog.java
@@ -2,7 +2,7 @@
  *
  * Pentaho Data Integration
  *
- * Copyright (C) 2002-2018 by Hitachi Vantara : http://www.pentaho.com
+ * Copyright (C) 2002-2023 by Hitachi Vantara : http://www.pentaho.com
  *
  *******************************************************************************
  *
@@ -56,6 +56,10 @@ import org.pentaho.di.ui.core.dialog.ErrorDialog;
 import org.pentaho.di.ui.core.widget.TextVar;
 import org.pentaho.di.ui.trans.step.BaseStepDialog;
 import org.pentaho.di.ui.trans.step.ComponentSelectionListener;
+import org.pentaho.di.core.annotations.PluginDialog;
+
+@PluginDialog( id = "CreditCardValidator", image = "CCV.svg", pluginType = PluginDialog.PluginType.STEP,
+        documentationUrl = "http://wiki.pentaho.com/display/EAI/Credit+card+validator" )
 
 public class CreditCardValidatorDialog extends BaseStepDialog implements StepDialogInterface {
   private static Class<?> PKG = CreditCardValidatorMeta.class; // for i18n purposes, needed by Translator2!!
