+++ plugins/credit-card-validator/ui/src/main/resources/CCV.png
